+++ src/app/features/owner/owner-detail/owner-detail.component.ts
@@ -2,6 +2,8 @@ import {Component, OnInit} from '@angular/core';
 import {ActivatedRoute, Router} from "@angular/router";
 import {OwnerService} from "../services/owner.service";
 import {OwnerBusinessModel} from "../../../common/business-models/owner.business-model";
+import {MatDialog} from "@angular/material/dialog";
+import {DialogComponent} from "../../../common/components/dialog/dialog.component";
 
 @Component({
   selector: 'app-owner-details',
@@ -16,7 +18,8 @@ export class OwnerDetailComponent implements OnInit {
   constructor(
     private activatedRoute: ActivatedRoute,
     private ownerService: OwnerService,
-    private router: Router
+    private router: Router,
+    private dialog: MatDialog
   ) {
   }
 
@@ -39,15 +42,24 @@ export class OwnerDetailComponent implements OnInit {
   }
 
   public deleteOwner(): void {
-    this.isLoading = true;
-    this.ownerService.deleteOwner(this.ownerId).subscribe({
-      next: () => {
-        this.isLoading = false;
-        this.router.navigate(["/proprietaires"]);
-      },
-      error: () => {
-        // TODO add Toast
-      }
+    const message = `Supprimer le propriétaire ${this.owner?.firstName} ${this.owner?.lastName} engendrera s'il y a, la
+    suppression du contrat associé et les paiements liés à ce contrat`
+    const dialogRef = this.dialog.open(DialogComponent, {
+      data: {message},
+      autoFocus: 'dialog'
+    });
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result === "cancel") return;
+      this.isLoading = true;
+      this.ownerService.deleteOwner(this.ownerId).subscribe({
+        next: () => {
+          this.isLoading = false;
+          this.router.navigate(["/proprietaires"]);
+        },
+        error: () => {
+          // TODO add Toast
+        }
+      })
     })
   }
 }
